Add files via upload
Water level controller is a simple yet effective way to prevent wastage of water. Its simplicity in
design and low cost components make it an ideal piece of technology for the common man.
One of the basic necessities of all living beings, water, is being wasted due to the uncontrolled
use and negligence of human beings. Therefore, it is important that we try and control the
wastage of water as much as we can.
Water level indicator is a simple electronic device that can help us save water by alarming us
when the water has reached the required level.